flask/requirements.txt
click==8.1.3
Flask==2.2.3
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
kafka-python==2.0.2
MarkupSafe==2.1.2
Werkzeug==2.2.3
zipp==3.14.0
elasticsearch==8.6.2
